Uses mel spectrogram on short non overlapping time segments to create (20) features per time segment. Uses k means to cluster these segments. The classes created are loosely reasonable on the test audio (is able to distinguish the presenter).
Also built a basic visualiser which plays audio and shows what cluster the current time segment belongs to to help with sense checks. It plays the audio externally with vlc though, with no attempt to resync, so there is the possibility of lag between the animated plot and the audio for lond audio samples.

Also produces a plot of the sound samples projected unto 3D (axes chosen to be the most significant a la PCA).

Too early to refactor.